fix about a thousand compiler warnings by doing s/u_char/char/ in a number
2007-11-02  John Harper  <jsh@unfactored.org>

	* values.c, lispmach.h, unix_processes.c, rep-md5.c, streams.c,
	message.c, lisp.c, numbers.c, rep-remote.c, gettext.c,
	regexp.c, unix_dl.c, rep_subrs.h, sockets.c, rep_lisp.h,
	unix_main.c, unix_files.c, rep_regexp.h, tables.c, lispcmds.c,
	find.c, main.c, symbols.c, misc.c: fix about a thousand
	compiler warnings by doing s/u_char/char/ in a number of places
	(in some other places cast back to u_char* to correctly convert
	characters to integers)


svn path=/trunk/; revision=2905

+++ src/main.c
@@ -183,7 +183,7 @@ check_configuration (int *stack_low)
     {
 	fprintf (stderr,
 	" ** error: --with-value-sizeof is incorrect; it should be %d\n",
-		 sizeof (rep_PTR_SIZED_INT));
+		 (int) sizeof (rep_PTR_SIZED_INT));
 	exit (10);
     }
